+++ packages/amplify-api-next/__tests__/ModelType.test.ts
@@ -392,4 +392,40 @@ describe('model auth rules', () => {
       });
     }
   }
+
+  it(`can define a custom authorization rule`, () => {
+    const schema = a.schema({
+      Widget: a.model({
+          title: a.string().required(),
+        })
+        .authorization([a.allow.custom()]),
+    });
+
+    const graphql = schema.transform().schema;
+    expect(graphql).toMatchSnapshot()
+  })
+
+  const TestOperations: Operation[][] = [
+    // each individual operation
+    ...Operations.map((op) => [op]),
+
+    // a couple sanity checks to support a combinations
+    ['create', 'read', 'update', 'delete'],
+    ['create', 'read', 'listen'],
+  ];
+
+  for (const operations of TestOperations) {
+    it(`can define custom auth rule for operations ${operations}`, () => {
+      const schema = a.schema({
+        widget: a
+          .model({
+            title: a.string().required(),
+          })
+          .authorization([a.allow.custom().to(operations)]),
+      });
+
+      const graphql = schema.transform().schema;
+      expect(graphql).toMatchSnapshot();
+    });
+  }
 });