+++ packages/amplify-api-next/src/MappedTypes/ModelMetadata.ts
@@ -24,6 +24,7 @@ export type ModelIdentifier<T> = {
 export type RelationalMetadata<
   ResolvedSchema,
   ResolvedFields extends Record<string, unknown>,
+  IdentifierMeta extends Record<string, any>,
 > = UnionToIntersection<
   ExcludeEmpty<
     {
@@ -54,7 +55,10 @@ export type RelationalMetadata<
                     Record<
                       // For M:N and 1:M we add a parent model field to the child
                       `${Uncapitalize<ModelName & string>}`,
-                      ResolvedFields[ModelName & string]
+                      NormalizeInputFields<
+                        ResolvedFields[ModelName & string],
+                        ExtractModelIdentifier<ModelName, IdentifierMeta>
+                      >
                     >
                   >;
                 }
@@ -63,7 +67,13 @@ export type RelationalMetadata<
                     Record<
                       // For 1:1 and Belongs To we add a child model field to the parent
                       Field,
-                      ResolvedFields[ResolvedSchema[ModelName][Field]['relatedModel']]
+                      NormalizeInputFields<
+                        ResolvedFields[ResolvedSchema[ModelName][Field]['relatedModel']],
+                        ExtractModelIdentifier<
+                          `${Capitalize<Field & string>}`,
+                          IdentifierMeta
+                        >
+                      >
                     >
                   >;
                 }
@@ -82,6 +92,15 @@ export type ExtractExplicitScalarFields<
     explicitScalarTypes: Scalars[ModelName];
   };
 };
+
+type ExtractModelIdentifier<ModelName, IdentifierMeta> =
+  ModelName extends keyof IdentifierMeta ? IdentifierMeta[ModelName] : never;
+
+type NormalizeInputFields<
+  ModelFields,
+  IdentifierMeta extends Record<string, any>,
+> = Partial<Omit<ModelFields, IdentifierMeta['identifier']>> &
+  Required<Pick<ModelFields, IdentifierMeta['identifier']>>;
 
 type ScalarFieldTypes<T> = {
   [ModelProp in keyof T]: {